# meson install perf improvements and new option --quiet