With these, `meson install` with gst-build on my Windows machine goes from taking 15.3 seconds to 4 seconds. `meson install --only-changed` goes from 11.6 seconds to 0.9 seconds.

Also fixes https://github.com/mesonbuild/meson/issues/4496.